requirements.txt
fastapi
uvicorn
jinja2
python-multipart
google-cloud-translate
google-auth
pypdf
pywebview
pyinstaller
